Move query history files to new folders
