+++ extensions/ql-vscode/src/query-history/query-history-info.ts
@@ -1,12 +1,12 @@
-import { RemoteQueryHistoryItem } from "./remote-queries/remote-query-history-item";
-import { VariantAnalysisHistoryItem } from "./remote-queries/variant-analysis-history-item";
-import { LocalQueryInfo } from "./query-results";
-import { assertNever } from "./pure/helpers-pure";
-import { pluralize } from "./pure/word";
+import { RemoteQueryHistoryItem } from "../remote-queries/remote-query-history-item";
+import { VariantAnalysisHistoryItem } from "./variant-analysis-history-item";
+import { LocalQueryInfo } from "../query-results";
+import { assertNever } from "../pure/helpers-pure";
+import { pluralize } from "../pure/word";
 import {
   hasRepoScanCompleted,
   getActionsWorkflowRunUrl as getVariantAnalysisActionsWorkflowRunUrl,
-} from "./remote-queries/shared/variant-analysis";
+} from "../remote-queries/shared/variant-analysis";
 
 export type QueryHistoryInfo =
   | LocalQueryInfo
