+++ extensions/ql-vscode/src/remote-queries/export-results.ts
@@ -13,7 +13,7 @@ import { Credentials } from "../authentication";
 import { ProgressCallback, UserCancellationException } from "../commandRunner";
 import { showInformationMessageWithAction } from "../helpers";
 import { extLogger } from "../common";
-import { QueryHistoryManager } from "../query-history";
+import { QueryHistoryManager } from "../query-history/query-history";
 import { createGist } from "./gh-api/gh-api-client";
 import { RemoteQueriesManager } from "./remote-queries-manager";
 import {
