+++ extensions/ql-vscode/src/remote-queries/remote-queries-view.ts
@@ -28,7 +28,7 @@ import {
 } from "./shared/remote-query-result";
 import { showAndLogWarningMessage } from "../helpers";
 import { URLSearchParams } from "url";
-import { SHOW_QUERY_TEXT_MSG } from "../query-history";
+import { SHOW_QUERY_TEXT_MSG } from "../query-history/query-history";
 import { AnalysesResultsManager } from "./analyses-results-manager";
 import { AnalysisResults } from "./shared/analysis-result";
 import { humanizeUnit } from "../pure/time";
