+++ extensions/ql-vscode/test/factories/query-history/local-query-history-item.ts
