+++ extensions/ql-vscode/test/factories/query-history/remote-query-history-item.ts
