+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-history/query-history-info.test.ts
@@ -1,20 +1,20 @@
-import { QueryStatus } from "../../../src/query-status";
+import { QueryStatus } from "../../../../src/query-status";
 import {
   buildRepoLabel,
   getActionsWorkflowRunUrl,
   getQueryId,
   getQueryText,
   getRawQueryName,
-} from "../../../src/query-history-info";
-import { VariantAnalysisHistoryItem } from "../../../src/remote-queries/variant-analysis-history-item";
-import { createMockVariantAnalysis } from "../../factories/remote-queries/shared/variant-analysis";
-import { createMockScannedRepos } from "../../factories/remote-queries/shared/scanned-repositories";
-import { createMockLocalQueryInfo } from "../../factories/local-queries/local-query-history-item";
-import { createMockRemoteQueryHistoryItem } from "../../factories/remote-queries/remote-query-history-item";
+} from "../../../../src/query-history/query-history-info";
+import { VariantAnalysisHistoryItem } from "../../../../src/query-history/variant-analysis-history-item";
+import { createMockVariantAnalysis } from "../../../factories/remote-queries/shared/variant-analysis";
+import { createMockScannedRepos } from "../../../factories/remote-queries/shared/scanned-repositories";
+import { createMockLocalQueryInfo } from "../../../factories/query-history/local-query-history-item";
+import { createMockRemoteQueryHistoryItem } from "../../../factories/query-history/remote-query-history-item";
 import {
   VariantAnalysisRepoStatus,
   VariantAnalysisStatus,
-} from "../../../src/remote-queries/shared/variant-analysis";
+} from "../../../../src/remote-queries/shared/variant-analysis";
 
 describe("Query history info", () => {
   const date = new Date("2022-01-01T00:00:00.000Z");
