+++ extensions/ql-vscode/test/vscode-tests/utils/query-history-helpers.ts
@@ -1,4 +1,4 @@
-import { QueryHistoryInfo } from "../../../src/query-history-info";
+import { QueryHistoryInfo } from "../../../src/query-history/query-history-info";
 
 export function shuffleHistoryItems(history: QueryHistoryInfo[]) {
   return history.sort(() => Math.random() - 0.5);
